增加登入狀態檢查的hook
