+++ next/pages/index.js
@@ -1,10 +1,13 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import MyApp from '@/pages/_app';
+import Head from 'next/head'
+import { Inter } from 'next/font/google'
+import styles from '@/styles/home.module.css'
 
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-  <React.StrictMode>
-    <MyApp />
-  </React.StrictMode>
-);
+const inter = Inter({ subsets: ['latin'] })
+
+export default function Home() {
+  return (
+    <>
+      <h1>假裝有的首頁</h1>
+    </>
+  )
+}
